+++ src/ecs/mod.rs
@@ -35,6 +35,7 @@ pub fn resolve_entity_icon(e: EntityRef<'_>) -> Option<char> {
     icon_from_component_panels!(
         // TODO(cohae): Custom havok icon
         // HavokShape,
+        Beacon,
         Ruler,
         Sphere,
         EntityModel,
@@ -66,7 +67,7 @@ pub fn resolve_entity_name(e: EntityRef<'_>, append_ent: bool) -> String {
             };
         }
 
-        name_from_component_panels!(Ruler, Sphere, EntityModel, StaticInstances);
+        name_from_component_panels!(Beacon, Ruler, Sphere, EntityModel, StaticInstances);
 
         format!("ent {}", e.entity().id())
     }